+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Unreleased
 - Add `OnErrorCallback::ignoreError` named constructor
 - [**BC**] Add `$initiator` argument to `ResponseDecoderInterface::supports` method
+- Copy `$response` in `ProfilerItem` not to store a reference to the original object
 
 ## 1.0.0 - 2021-05-12
 - Initial implementation